Fix script/setup on initial run
Why is this change needed?
--------------------------
On the first run we install homebrew, but we don't load it into the
shell. This causes the script to fail when it tries to use `brew`.

How does it address the issue?
------------------------------
Setup homebrew after installing it.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?