Improve setup_secrets
Why is this change needed?
--------------------------
The previous implimentation stored the path of the image in
`Dropbox/.dropbox_secrets_path_$(hostname -s)` which annoyingly wasn't
visible in the dropbox.com UI. We move it to a more visible location.

We also collect everything under `Documents/secrets/$(hostname -s)` to
keep things organized instead of having many files in the dropbox root.

I was curious about moving more stuff into the secrets volume so I
chatted with GPT about the sparse disk options and it mentioned
sparsebundles having better network reslience. I'll give it a shot and
see.

I also moved the secrets volume to iCloud `Documents` because so it
should automatically sync to a new laptop when I log in with my iCloud
account.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?