Fix brewfile issue
Why is this change needed?
--------------------------
`File.dirname(__FILE__)` was returning `~/.dotfiles` instead of the
user's home directory. This cause us to not find user brewfiles that
were linked through stow.

How does it address the issue?
------------------------------
Look for `.brewfiles` in the user's home directory.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?