Fix stow regex issues
Why is this change needed?
--------------------------
The patterns in this file are interpreted as perl regexes, so we need to
escape special characters. Otherwise stow will throw an error for
filenames that have regex characters in them.

How does it address the issue?
------------------------------
Escape special characters in the regexes.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?