Fix stow tree folding for completely masked dirs
Why is this change needed?
--------------------------
If I add a new ssh key, I want it to show up in my secrets volume.

How does it address the issue?
------------------------------
Ignore dirs in stow packages that are completely masked by more
important packages. This results in ~/.ssh being a symlink to
/Volumes/secrets/.ssh, instead of a directory with symlinks inside of
it.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?