Remove .common_files_remote_jason
Why is this change needed?
--------------------------
Bastion MFA stopped this from working, and I don't have time to fix it
at the moment.

How does it address the issue?
------------------------------
Remove this dir so we don't try to sync these files to a remote host.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?