Move secrets to Dropbox
Why is this change needed?
--------------------------
In my case Finder always indicates that iCloud is in the process of
syncing, usually somewhere between 10%-50% done. This seems equivalent
to not syncing for practical purposes.

Matt reports having issues with iCloud eating files.

How does it address the issue?
------------------------------
Store the volume in Dropbox, which seems more reliable.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://github.com/expectedbehavior/common-files/pull/55

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?