Fix issue with iTerm and stow
Why is this change needed?
--------------------------
Stow can't handle absolute symlinks yet, which causes an error if trying
to setup stow stuff after iTerm has created this symlink.

How does it address the issue?
------------------------------
Ignore the symlink, which means stow links the containing directory,
which results in what we want (iTerm creating the symlink in the
secrets dir).

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?