Update all Brewfiles
Why is this change needed?
--------------------------
It's been a long time since it was reviewed and there was a lot of cruft. Some things were missing that we use and lots of things were included that we no longer use.

How does it address the issue?
------------------------------
The following kinds of changes were made
- The categorization changed to make it easier to comply. That's reflected by the file names and comments
- In each file, the software is alphabetized
- I added a personal Brewfile for Jason (with his consent) and updated my own to move things that are just for us into the right place

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------

Any screenshots?
----------------

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?

Co-authored-by: Jason Gladish <jason@expectedbehavior.com>