Add rbspy
Why is this change needed?
--------------------------
This is handy for seeing what's going on with a running ruby process,
which is very handy for debugging.

How does it address the issue?
------------------------------
Install rbspy via homebrew.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
https://rbspy.github.io/

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?