Add Firefox ESR to the path if it is installed
Why is this change needed?
--------------------------
We use Firefox ESR in some projects because it works with the latest version of geckodrivers[1]

How does it address the issue?
------------------------------
Check to see if it is installed at the place that Homebrew usually install it. If it is installed, add it to the path.

Any links to any relevant tickets, articles, or other resources?
---------------------------------------------------------------
1 - https://firefox-source-docs.mozilla.org/testing/geckodriver/Support.html

Did you complete all of the following?
--------------------------------------
- Run test suite? -
- Add new tests? -
- Consider security implications and practices? Y