Make lc work with automation notifications
Why is this change needed?
--------------------------
Automation notifications use a different group name, but I want to be
able to clear them like I do with other notifications.

How does it address the issue?
------------------------------
Search for `automation-notify` notifications with terminal-notifier.

This also reformats a bit because the line was getting long.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?