Update karabiner config
Why is this change needed?
--------------------------
I think this fixed some key combo that wasn't working. Maybe fn/globe?

How does it address the issue?
------------------------------
It has fallen out of my head, but I think it was necessary and very
annoying, so I'm not excited about figuring it out again.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?