Make delta show hunk line numbers
Why is this change needed?
--------------------------
This is super helpful because it allows cmd-clicking in iterm files
directly to the line in the file (or close to).

How does it address the issue?
------------------------------
Set the hunk formatting and some unique colors to make it more readable.

Did you complete all of the following?
--------------------------------------
- Run test suite?
- Add new tests?
- Consider security implications and practices?